feat(kernel): Introduce KernelInputSpec for input validation and enhance KernelContract
- Added KernelInputSpec struct to define input keys and their expected counts.
- Updated KernelContract to utilize KernelInputSpec for input validation.
- Enhanced validation logic in KernelContract to ensure correct input layout.
- Modified KernelRegistry to check input layout against KernelContract during execution.
- Improved error handling with new error types for input layout mismatches and output contract mismatches.
- Updated tests to cover new input validation logic and error scenarios.
- Refactored RuntimeValue to use Arc for lists and records for better memory management.
- Added new error messages for improved user feedback in case of input/output mismatches.

+++ src-tauri/crates/yss-application/src/ipc/schema/editor_projection.rs
@@ -545,7 +545,7 @@ mod tests {
             wire["nodes"][0]["ports"][0]["resolvedSchema"],
             json!({
                 "kind": "derived",
-                "fields": [{"name": "sales", "scalarType": "float64"}],
+                "fields": [{"name": "sales", "scalarType": "Numeric"}],
             })
         );
         assert_eq!(wire["nodes"][0]["portInstanceAdditions"], json!([]));
@@ -555,7 +555,7 @@ mod tests {
                 "kind": "projectColumns",
                 "available": true,
                 "unavailableReason": null,
-                "options": [{"name": "sales", "dataType": "float64"}],
+                "options": [{"name": "sales", "dataType": "Numeric"}],
                 "value": ["sales"],
             })
         );
